+++ src/api/java/com/codetaylor/mc/artisanworktables/api/package-info.java
@@ -1,4 +1,4 @@
-@API(owner = "artisanworktables", apiVersion = "9", provides = "ArtisanWorktablesAPI")
+@API(owner = "artisanworktables", apiVersion = "10", provides = "ArtisanWorktablesAPI")
 package com.codetaylor.mc.artisanworktables.api;
 
 import net.minecraftforge.fml.common.API;