+++ src/api/java/com/codetaylor/mc/artisanworktables/api/pattern/IItemDesignPattern.java
@@ -1,0 +1,15 @@
+package com.codetaylor.mc.artisanworktables.api.pattern;
+
+import net.minecraft.item.ItemStack;
+
+import javax.annotation.Nullable;
+
+public interface IItemDesignPattern {
+
+  boolean hasRecipe(ItemStack itemStack);
+
+  @Nullable
+  String getRecipeName(ItemStack itemStack);
+
+  void setRecipeName(ItemStack itemStack, String recipeName);
+}