+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/item/ItemDesignPattern.java
@@ -1,31 +1,37 @@
 package com.codetaylor.mc.artisanworktables.modules.worktables.item;
 
+import com.codetaylor.mc.artisanworktables.api.pattern.IItemDesignPattern;
 import com.codetaylor.mc.artisanworktables.modules.worktables.ModuleWorktables;
 import com.codetaylor.mc.artisanworktables.modules.worktables.ModuleWorktablesConfig;
+import com.codetaylor.mc.athenaeum.util.StackHelper;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumHand;
 import net.minecraft.world.World;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
-@SuppressWarnings("WeakerAccess")
 public class ItemDesignPattern
-    extends Item {
+    extends Item
+    implements IItemDesignPattern {
 
   public static final String NAME = "design_pattern";
   public static final String NAME_BLANK = "blank";
   public static final String NAME_WRITTEN = "written";
 
+  private static final String TAG_ARTISAN_WORKTABLES = "ArtisanWorktables";
+  private static final String TAG_RECIPE_NAME = "RecipeName";
+
   @Nonnull
   @Override
   public String getUnlocalizedName(ItemStack stack) {
 
-    // TODO: fix this
-    if (stack.hasTagCompound()) {
+    if (this.hasRecipe(stack)) {
       return super.getUnlocalizedName(stack) + "." + NAME_WRITTEN;
     }
 
@@ -44,5 +50,41 @@ public class ItemDesignPattern
     }
 
     return new ActionResult<>(EnumActionResult.PASS, player.getHeldItem(hand));
+  }
+
+  @Override
+  public boolean hasRecipe(ItemStack itemStack) {
+
+    return (this.getRecipeName(itemStack) != null);
+  }
+
+  @Nullable
+  @Override
+  public String getRecipeName(ItemStack itemStack) {
+
+    if (itemStack.getItem() == this) {
+      NBTTagCompound tagCompound = StackHelper.getTagSafe(itemStack);
+
+      if (tagCompound.hasKey(TAG_ARTISAN_WORKTABLES)) {
+        NBTTagCompound innerTagCompound = tagCompound.getCompoundTag(TAG_ARTISAN_WORKTABLES);
+
+        if (innerTagCompound.hasKey(TAG_RECIPE_NAME)) {
+          return innerTagCompound.getString(TAG_RECIPE_NAME);
+        }
+      }
+    }
+
+    return null;
+  }
+
+  @Override
+  public void setRecipeName(ItemStack itemStack, String recipeName) {
+
+    if (itemStack.getItem() == this) {
+      NBTTagCompound tagCompound = StackHelper.getTagSafe(itemStack);
+      NBTTagCompound innerTagCompound = tagCompound.getCompoundTag(TAG_ARTISAN_WORKTABLES);
+      innerTagCompound.setString(TAG_RECIPE_NAME, recipeName);
+      tagCompound.setTag(TAG_ARTISAN_WORKTABLES, innerTagCompound);
+    }
   }
 }