+++ docs/changelog.md
@@ -5,6 +5,10 @@
         * `tile.artisanworktables.automator.name`
         * `tile.artisanworktables.automator_power_rf.name`
         * `gui.artisanworktables.tooltip.automator.power`
+        
+  * **API:**
+    * Bumped version to 10
+    * Added `com.codetaylor.mc.artisanworktables.api.pattern.IItemPattern`
 
   * **Added:**
     * Recipe id now displays on the bottom of the tooltip for JEI's output item [(#200)](https://github.com/codetaylor/artisan-worktables/issues/200)